Update client.lua
Updated to version 2.00.

Updated client side functions for notifying 911. Defaults to send calls to Night's nightshift MDT system. Specific section for servers to edit the 911 function to fit the needs of their 911 system if triggerred client side. (This event is triggerred by a Server Event so if your 911 uses server side exports/events to trigger, that can be done in the server.lua.

Added functionality to allow for random length of robberies vs fixed length using min/max times set in config.